Refactor the code for readability, use C++
