+++ Makefile
@@ -1,6 +1,5 @@
 XBE_TITLE = xblunblock
 GEN_XISO = $(XBE_TITLE).iso
 
-SRCS +=	main.c
+SRCS += main.cpp
 include $(NXDK_DIR)/Makefile
-
